test(connection): extract common.rs tests to sibling file
Per CLAUDE.md, tests must live in their own *_tests.rs file wired in
via `#[path = "..."] mod tests;`, not in inline `#[cfg(test)] mod tests`
blocks. `connection/common.rs` was the lone holdout; sync.rs and
async.rs already follow the convention. Move every test (including the
new require_protobuf_support tests) into common_tests.rs, no logic
changes.
